refactor: add helpers module
